add clients and related

+++ database/migrations/user/0001_00_00_000001_create_users_table.php
@@ -16,7 +16,7 @@ class CreateUsersTable extends Migration
     {
         Schema::create('users', static function (Blueprint $table) {
 
-            $table->unsignedInteger('id', true);
+            $table->unsignedBigInteger('id', true);
 
             $table->string('login')->unique()->nullable();
             $table->string('password')->nullable();
