+++ database/migrations/user/0001_00_00_000002_create_user_profiles_table.php
@@ -15,7 +15,7 @@ class CreateUserProfilesTable extends Migration
     {
         Schema::create('user_profiles', static function (Blueprint $table) {
 
-            $table->unsignedInteger('user_id')->unique()->primary();
+            $table->unsignedBigInteger('user_id')->unique()->primary();
 
             $table->string('lastname');
             $table->string('firstname');
